+++ rmpd/src/app.rs
@@ -154,7 +154,7 @@ async fn restore_state(
                         status.audio_format = Some(rmpd_core::song::AudioFormat {
                             sample_rate: sr,
                             channels: ch,
-                            bits_per_sample: bps,
+                            bits_per_sample: bps as u8,
                         });
                     }
                     drop(status);
